refactor: Login Finished Page
