+++ frontend-hybrid/yakal/lib/main.dart
@@ -8,10 +8,10 @@ import 'package:yakal/screens/Login/Identification/screen.dart';
 import 'package:yakal/screens/Login/IdentificationEntry/screen.dart';
 import 'package:yakal/screens/Login/IdentificationResult/screen.dart';
 import 'package:yakal/screens/Login/LoginEntry/screen.dart';
+import 'package:yakal/screens/Login/LoginFinished/screen.dart';
 import 'package:yakal/screens/Login/LoginTerms/screen.dart';
 import 'package:yakal/screens/Login/ModeSelection/screen.dart';
 import 'package:yakal/screens/Login/NicknameInput/screen.dart';
-import 'package:yakal/screens/Login/auth_finish_screen.dart';
 import 'package:yakal/screens/Profile/profile_screen.dart';
 import 'package:yakal/screens/Setting/setting_screen.dart';
 import 'package:yakal/screens/Setting/setting_signout_screen.dart';
@@ -84,7 +84,7 @@ class MyApp extends StatelessWidget {
             ),
             GetPage(
               name: '/finish',
-              page: () => const AuthFinishScreen(),
+              page: () => const LoginFinishedScreen(),
             ),
           ],
         ),
