+++ frontend-hybrid/yakal/lib/screens/Login/LoginFinished/style.dart
@@ -1,0 +1,19 @@
+import 'package:flutter/material.dart';
+
+import '../../../utilities/style/color_styles.dart';
+
+class LoginFinishedStyle {
+  static const name = TextStyle(
+    color: ColorStyles.black,
+    fontSize: 24,
+    fontWeight: FontWeight.w700,
+    height: 1.4,
+  );
+
+  static const title = TextStyle(
+    color: ColorStyles.black,
+    fontSize: 24,
+    fontWeight: FontWeight.w600,
+    height: 1.4,
+  );
+}
